Improve overrides of consult-projectile-sources
This allows using consult-projectile with
consult-projectile-switch-project, like this:

(setq consult-projectile-source-projectile-project-action
      (lambda (dir)
        (let ((default-directory dir))
          (consult-projectile))))

The previous implementation (using a let bind) affected
consult-projectile-sources for recursive calls. In the snippet above,
this meant that consult-projectile-switch-project would get stuck in a
loop selecting a project repeatedly instead of selecting a project then
using the original consult-projectile-sources to select a
file/buffer/etc.

The new implementation uses an optional argument so that only the
initial/outer call to consult-projectile is affected.